pyproject.toml
[project]
name = "local-transcriber"
version = "0.1.0"
description = "Local CLI tool for audio/video transcription using faster-whisper"
requires-python = ">=3.10"
dependencies = [
    "typer",
    "rich",
    "faster-whisper>=1.2.1",
    "socksio>=1.0.0",
    "nvidia-cublas-cu12>=12.4; sys_platform == 'linux' and platform_machine == 'x86_64'",
]

[project.scripts]
transcribe = "local_transcriber.cli:app"

[dependency-groups]
dev = [
    "pytest",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["src/local_transcriber"]
